Add Assurance support to Android Sample App
Add Assurance support to Android Sample App
