+++ samples/launch/mobile/android/BasicPlayerSample/src/main/java/com/adobe/mediaanalyticstestapp/player/ObservableVideoView.java
@@ -2,7 +2,7 @@
  * ADOBE CONFIDENTIAL
  * ___________________
  *
- * Copyright 2018 Adobe
+ * Copyright 2021 Adobe
  * All Rights Reserved.
  *
  * NOTICE: All information contained herein is, and remains
