+++ samples/launch/mobile/android/BasicPlayerSample/src/main/res/layout/activity_main.xml
@@ -4,25 +4,32 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <com.adobe.mediaanalyticstestapp.player.ObservableVideoView
-        android:id="@+id/videoView"
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent"
-        android:layout_centerHorizontal="true"
-        android:layout_centerVertical="true" />
-
-    <TextView
-        android:id="@+id/adOverlayView"
-        android:layout_width="50dp"
-        android:layout_height="50dp"
+    <Button
+        android:id="@+id/assuranceButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="30dp"
-        android:layout_marginTop="28dp"
-        android:background="@android:color/background_dark"
-        android:gravity="center"
-        android:text="@string/ad_overlay_text"
-        android:textAppearance="?android:attr/textAppearanceLarge"
-        android:textColor="@android:color/white"
-        android:visibility="invisible" />
+        android:layout_marginTop="100dp"
+        android:onClick="setAssurance"
+        android:text="Set Assurance"/>
+
+
+    <EditText
+        android:id="@+id/assuranceUrl"
+        android:layout_width="fill_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="20dp"
+        android:gravity="left"
+        android:cursorVisible="true"
+        android:hint="Enter Assurance Url" />
+
+
+    <Button
+        android:id="@+id/startVideoPlayer"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true"
+        android:text="Start Video Player"
+        android:textStyle="bold"/>
 </RelativeLayout>
